--- dfir_rs/examples/example_naturals.rs.disabled
@@ -1,23 +1,0 @@
-use dfir_rs::dfir_syntax;
-
-pub fn main() {
-    let mut _flow = dfir_syntax! {
-        base = source_iter(vec![1]) -> cycle;
-        cycle = union()
-                -> map(|i| i + 1)
-                -> inspect(|i| println!("{}", i))
-                -> cycle;
-    };
-
-    // Let's not run this -- it will go forever!
-    // flow.run_available_sync();
-}
-
-#[cfg(not(nightly))]
-#[test]
-fn test() {
-    use example_test::run_current_example;
-
-    let output = run_current_example!().read_to_end();
-    hydro_build_utils::assert_snapshot!("example_naturals", output);
-}
