+++ dfir_rs/tests/surface_cross_join_multiset.rs
@@ -2,7 +2,7 @@ use std::cell::RefCell;
 use std::collections::HashMap;
 use std::rc::Rc;
 
-use dfir_rs::scheduled::ticks::TickInstant;
+use dfir_rs::runtime::ticks::TickInstant;
 use multiplatform_test::multiplatform_test;
 
 macro_rules! assert_contains_each_by_tick {
