+++ dfir_rs/tests/surface_fold_keyed.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeSet;
 
 use dfir_rs::assert_graphvis_snapshots;
-use dfir_rs::scheduled::ticks::TickInstant;
+use dfir_rs::runtime::ticks::TickInstant;
 use dfir_rs::util::collect_ready;
 use multiplatform_test::multiplatform_test;
