+++ dfir_rs/tests/surface_persist.rs
@@ -1,6 +1,6 @@
 use dfir_pipes::pull::HalfMultisetJoinState;
 use dfir_rs::assert_graphvis_snapshots;
-use dfir_rs::scheduled::ticks::TickInstant;
+use dfir_rs::runtime::ticks::TickInstant;
 use dfir_rs::util::collect_ready;
 use multiplatform_test::multiplatform_test;
