+++ dfir_rs/tests/surface_singleton.rs
@@ -1,5 +1,5 @@
 use dfir_rs::assert_graphvis_snapshots;
-use dfir_rs::scheduled::ticks::TickInstant;
+use dfir_rs::runtime::ticks::TickInstant;
 use dfir_rs::util::collect_ready;
 use lattices::Max;
 use multiplatform_test::multiplatform_test;
