+++ dfir_rs/tests/surface_state_scheduling.rs
@@ -1,5 +1,5 @@
 use dfir_rs::dfir_syntax;
-use dfir_rs::scheduled::ticks::TickInstant;
+use dfir_rs::runtime::ticks::TickInstant;
 use dfir_rs::util::collect_ready;
 use multiplatform_test::multiplatform_test;
